Introduce useSignAndSendTransaction hook
