+++ packages/ui/src/modals/AddMembershipModal/SignCreateMemberModal.tsx
@@ -1,9 +1,7 @@
-import { SubmittableExtrinsic } from '@polkadot/api/types'
-import { web3FromAddress } from '@polkadot/extension-dapp'
 import { BalanceOf } from '@polkadot/types/interfaces/runtime'
 import { ISubmittableResult } from '@polkadot/types/types'
 import BN from 'bn.js'
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import { Observable } from 'rxjs'
 import { ButtonPrimaryMedium } from '../../components/buttons'
 import { Label } from '../../components/forms'
@@ -11,10 +9,8 @@ import { Help } from '../../components/Help'
 import { Modal, ModalBody, ModalFooter, ModalHeader } from '../../components/Modal'
 import { SelectAccount } from '../../components/selects/SelectAccount'
 import { Text, TokenValue } from '../../components/typography'
-import { Account } from '../../hooks/types'
 import { useApi } from '../../hooks/useApi'
-import { useKeyring } from '../../hooks/useKeyring'
-import { useObservable } from '../../hooks/useObservable'
+import { useSignAndSendTransaction } from '../../hooks/useSignAndSendTransaction'
 import { BalanceInfoNarrow, InfoTitle, InfoValue, Row } from '../common'
 import { Params } from './MembershipFormModal'
 
@@ -25,37 +21,7 @@ interface SignProps {
   onSign: (transaction: Observable<ISubmittableResult>, fee: BN) => void
 }
 
-export function useSignAndSendTransaction(
-  transfer: SubmittableExtrinsic<'rxjs'> | undefined,
-  from: Account,
-  onSign: (transaction: Observable<ISubmittableResult>, fee: BN) => void
-) {
-  const [isSending, setIsSending] = useState(false)
-  const keyring = useKeyring()
-  const { api } = useApi()
-  const paymentInfo = useObservable(transfer?.paymentInfo(from.address), [from])
-
-  useEffect(() => {
-    if (!isSending || !transfer || !paymentInfo) {
-      return
-    }
-
-    const keyringPair = keyring.getPair(from.address)
-    const fee = paymentInfo.partialFee.toBn()
-
-    if (keyringPair.meta.isInjected) {
-      web3FromAddress(from.address).then(({ signer }) => {
-        onSign(transfer.signAndSend(from.address, { signer: signer }), fee)
-      })
-    } else {
-      onSign(transfer.signAndSend(keyringPair), fee)
-    }
-  }, [api, isSending])
-
-  return { isSending, send: () => setIsSending(true), paymentInfo }
-}
-
-export const SignTransactionModal = ({ onClose, membershipPrice, transactionParams, onSign }: SignProps) => {
+export const SignCreateMemberModal = ({ onClose, membershipPrice, transactionParams, onSign }: SignProps) => {
   const { api } = useApi()
   const [from, setFrom] = useState(transactionParams.controllerAccount)
   const transfer = api?.tx?.members?.buyMembership({
@@ -66,8 +32,7 @@ export const SignTransactionModal = ({ onClose, membershipPrice, transactionPara
     avatar_uri: transactionParams.avatar,
     about: transactionParams.about,
   })
-
-  const { paymentInfo, isSending, send } = useSignAndSendTransaction(transfer, from, onSign)
+  const { paymentInfo, isSending, send } = useSignAndSendTransaction({ transfer, from, onSign })
 
   return (
     <Modal modalSize="m" modalHeight="s">
